Update RE_EDS_general_analysis.py (#53)
I fixed a bug where the undersampling index provided by the `undersampling_detection` and `sampling_analysis` parts of the code used an indexing convention where the first replica was indexed as 1. Whereas the code in the energy offsets uses an index of 0. 

The `sampling_results` dictionary is currently used in more places in the code which is why I did not modify how it is generated directly. But in the future it might be preferable to rewrite some of the code to ensure the same convention is used throughout.